Refactored API and explorer modules, added API methods.
