+++ nomen/src/subcommands/server/api.rs
@@ -2,9 +2,11 @@ use axum::{
     extract::{Query, State},
     Json,
 };
-use nomen_core::CreateBuilder;
+use nomen_core::{CreateBuilder, TransferBuilder};
 
 use crate::db;
+
+use self::models::{OpReturnResponse, TransferEventResponse};
 
 use super::{AppState, WebError};
 
@@ -13,7 +15,8 @@ mod models {
 
     use askama_axum::IntoResponse;
     use axum::{http::StatusCode, Json};
-    use secp256k1::XOnlyPublicKey;
+    use nostr_sdk::UnsignedEvent;
+    use secp256k1::{schnorr::Signature, XOnlyPublicKey};
     use serde::{Deserialize, Serialize};
 
     #[derive(Serialize)]
@@ -67,6 +70,50 @@ mod models {
     pub struct OpReturnResponse {
         pub op_return: Vec<String>,
     }
+
+    #[derive(Deserialize)]
+    pub struct TransferEventQuery {
+        pub name: String,
+        pub new_owner: XOnlyPublicKey,
+        pub old_owner: XOnlyPublicKey,
+    }
+
+    #[derive(Serialize)]
+    pub struct TransferEventResponse {
+        pub event: UnsignedEvent,
+    }
+
+    #[derive(Deserialize)]
+    pub struct TransferQuery {
+        pub name: String,
+        pub new_owner: XOnlyPublicKey,
+        pub signature: Signature,
+    }
+
+    #[derive(Serialize)]
+    pub struct NameResponse {
+        pub name: String,
+        pub pubkey: String,
+    }
+
+    #[derive(Serialize)]
+    pub struct NamesResponse {
+        pub names: Vec<NameResponse>,
+    }
+}
+
+pub async fn names(
+    State(state): State<AppState>,
+) -> Result<Json<models::NamesResponse>, models::JsonError> {
+    let names = db::all_names(&state.pool)
+        .await?
+        .into_iter()
+        .map(|(n, pk)| models::NameResponse {
+            name: n,
+            pubkey: pk,
+        })
+        .collect();
+    Ok(Json(models::NamesResponse { names }))
 }
 
 pub async fn name(
@@ -93,4 +140,32 @@ pub async fn op_return_v1(
     };
 
     Ok(Json(orr))
+}
+
+#[allow(clippy::unused_async)]
+pub async fn get_transfer_event(
+    Query(query): Query<models::TransferEventQuery>,
+) -> Result<Json<models::TransferEventResponse>, models::JsonError> {
+    let tb = TransferBuilder {
+        new_pubkey: &query.new_owner,
+        name: &query.name,
+    };
+    Ok(Json(TransferEventResponse {
+        event: tb.unsigned_event(&query.old_owner),
+    }))
+}
+
+#[allow(clippy::unused_async)]
+pub async fn get_transfer(
+    Query(query): Query<models::TransferQuery>,
+) -> Result<Json<models::OpReturnResponse>, models::JsonError> {
+    let tb = TransferBuilder {
+        new_pubkey: &query.new_owner,
+        name: &query.name,
+    };
+    let or1 = hex::encode(tb.transfer_op_return());
+    let or2 = hex::encode(tb.signature_provided_op_return(query.signature));
+    Ok(Json(OpReturnResponse {
+        op_return: vec![or1, or2],
+    }))
 }
